fix: run ocr alone without opencv
