+++ examples/uitest/demo_weixin_live.yaml
@@ -33,17 +33,20 @@ teststeps:
         actions:
             - method: swipe_to_tap_text
               params: 轻触进入直播间
+              identifier: 进入直播间
               max_retry_times: 10
     - name: 向上滑动，等待 10s
       ios:
         actions:
             - method: swipe
               params: up
+              identifier: 第一次上划
             - method: sleep
               params: 10
             - method: screenshot
             - method: swipe
               params: up
+              identifier: 第二次上划
             - method: sleep
               params: 10
             - method: screenshot
